style: switch to Jekyll with 'Just the Docs' theme
To prepare a rework of the website with more content, Jekyll [1] is now
used with the 'Just the Docs' theme [2]. This also includes a GitHub
Action workflow to do the build and deployment automatically when the
code is modified.

A few modifications have been done:
- nav.html: Also list the sub-titles from the different pages
- custom.js: switch to dark theme automatically
- custom.scss: 'justify' alignment, 'click to see more', colours
- mermaid is loaded with FA support: it will be used to generate schemas

It also includes a new logo under CC 0 license, by Jonas Gendebien.

A big thank you to Maxime for having conducted this switch to Jekyll.

Link: https://jekyllrb.com/ [1]
Link: https://just-the-docs.com [2]
Co-developed-by: Maxime Dourov (@mux99) <m.dourov@students.ephec.be>
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>