pm: describe the behavior
I think this was missing, and leading to confusions.

With mptcp.org, the 'fullmesh' PM was often used by default: it is
easier to explain, but probably not the best as it was creating too many
subflows. (Even if new subflows can be blocked with netfilter or ip
rules.)

Here, different cases have to be taken into account, but hopefully that
is clear enough.

Co-authored-by: Mat Martineau <16723289+mjmartineau@users.noreply.github.com>
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>