apps: add 'Linux distributions' section
As pointed out by Arınç and Daniel [1], it would be clearer to have
a dedicated Linux distributions sections, because OpenWrt is not an
application.

Here I tried to list most common GNU/Linux distributions, avoiding forks
having the same kernel configuration and mptcpd package. Raspbian and
Ubuntu have different kernel configurations and are quite popular, it is
better to mention them.

I decided to add a new section in the 'apps' page, not to create yet
another page, while it is linked to the rest.

Because it is better not to rename a page to avoid breaking links, I
changed the title instead: Support in apps / OS. It covers what is
currently in this page, and still linked to 'apps'.

Link: https://github.com/multipath-tcp/mptcp.dev/pull/26 [1]
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>