pm: NM >= 1.40 set the endpoints by default
It is important to know that, not to have NM and mptcpd configuring the
endpoints. Also, in most cases, it might not be needed to configure
these endpoints if NM or mptcpd does the job for the user.

While at it, mention that the in-kernel PM is the default one, and add
a note about `rp_filter`, the same as the one mentioned in NM's release
notes.

Link: https://networkmanager.dev/blog/networkmanager-1-40/#mptcp-support
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>