faq: enable MPTCP support with OpenSSH
Because OpenSSH maintainers do not want to add a native MPTCP support,
it is not that obvious what to do to enable this support: on the server
side, the socket might be created by systemd, while on the client side,
GUI might use SSH, but they are not launched from a terminal.

A few different cases are now covered here.

Because the list is a bit long, it will not be displayed by default, a
'(click to see more)' label will be visible instead.

Reviewed-by: Mat Martineau <martineau@kernel.org>
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>